create new branch and rename screens
